Add session results page and pronunciation feedback components
- Introduced `SessionResultsPage` for displaying results from solo and group sessions.
- Added `PronunciationFeedbackMock`, `WordFeedback`, and `PhonemeFeedback` models for handling pronunciation assessment data.
- Implemented interactive feedback widgets including `InteractiveFeedbackSentence` and `PronunciationMetricColumn` for visualizing pronunciation scores.
- Created `SessionBreakdownTile` to show detailed exercise summaries with phoneme-level feedback.
- Updated routing to navigate to the new results page after group sessions.
- Refactored pronunciation feedback imports to a common model for consistency.
