+++ apps/mobile_interface/lib/src/features/group_session/pages/group_session_active_lobby_page.dart
@@ -172,7 +172,7 @@ class _GroupSessionActiveLobbyPageState extends State<GroupSessionActiveLobbyPag
 
     Navigator.pushReplacementNamed(
       context,
-      AppRoutes.groupSessionPostSession,
+      AppRoutes.groupSessionResults,
       arguments: participants,
     );
   }
