+++ apps/mobile_interface/lib/src/features/solo_practice/services/pronunciation_feedback_service.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:http/http.dart' as http;
 
-import '../models/pronunciation_feedback.dart';
+import '../../../common/models/pronunciation_feedback.dart';
 
 /// API gateway base URL used for local development.
 ///
